Added Binary Adder Turing Machine spec and tapes
The binadd.tm file contains the specification for a Turing Machine that
can verify a tape of three binary strings separated by hashes such that
if the first two binary strings add up to the third, the machine
accepts, and if they do not it rejects.

Also included are a suite of pre-generated tapes and bash scripts that
can both generate and run the machine for a variety of complexities of
input.